Delete the existing tests
These tests are unfortunately all testing functionality that is set to
be removed in the scope of this work. They will be replaced with new—and
relevant—tests before this work item is complete.

This commit also updates the foundry configuration with some useful
tweaks to improve the debugging experience.