Use supersede time, not add time, for root timestamps
Previously, roots were getting a timestamp associated with them that was
the time they were first put on chain. This meant that, if there was a
period of no root updates, it would be possible for a root change to
invalidate all historical roots and hence invalidate all in-flight
transactions at the time.

Now we associate the root with the timestamp at which they were
superseded. In other words, the timestamp for the block in which they
were added to the history.