Validate inputs for being in reduced form
We now check the provided `identityCommitments` for being in reduced
form, reverting if they are not. We also check `preRoot` and `postRoot`
for being in this form with the same action on failure.

This commit adds a descriptive `UnreducedElement` error type with an
enum `UnreducedValueType` that describes the value accompanying it in
the error message.