Add a low-level test for type safety violations
As we cannot directly test whether providing an out-of-range value for
startIndex is safe due to Solidity's type safety, we instead call into
the function outside of the regular flow using the low-level call
operations. This is sufficient proof that we do not need an additional
check on that input, and can hence save gas.