Update the deploy scripts to handle upgrades (#40)
The deploy scripts are now capable of upgrading the WorldID identity
manager contract in place on chain. This is not built to be used for
production deploys, but instead to act as documentation, and to allow
upgrades during testing for the signup-sequencer.

It also adds support for saving configuration to the deploy scripts,
thereby making it easier to run them repeatedly when working on the
signup sequencer.