Allow runs of zeroes for identities (#47)
Previously, an identity of zero anywhere in the array of identity
commitments was considered to be a reason to fail. However, it turns out
that we want to be able to submit batches that are padded out to the
batch size using zeroes.

This commit changes the validation to only consider a non-zero identity 
occurring after a zero identity in the array as invalid.